+++ qlty-cli/src/ui/messages.rs
@@ -1,4 +1,5 @@
 use anyhow::Result;
+use console::style;
 use qlty_check::Report;
 use qlty_types::analysis::v1::Message;
 
@@ -9,13 +10,18 @@ pub fn print_installation_error_messages(
     let installation_error_messages: Vec<&Message> = report
         .messages
         .iter()
-        .filter(|m| m.ty == "executor.install.error")
+        .filter(|m| m.ty.starts_with("executor.install."))
         .collect();
 
     if !installation_error_messages.is_empty() {
-        writeln!(writer, "Installation errors found:")?;
+        writeln!(writer)?;
+        writeln!(
+            writer,
+            "{}",
+            style("Installation errors found:").red().bold()
+        )?;
         for message in &installation_error_messages {
-            writeln!(writer, "{}", message.message)?;
+            writeln!(writer, "  - {}", message.message)?;
         }
         writeln!(writer)?;
     }